requirements.txt
altair==5.3.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
protobuf==4.25.3
pure-eval==0.2.2
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.35.0
streamlit-server-state==0.17.1
tenacity==8.3.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
